refactor(navi-menu): rename to navi-catalog
