+++ src/components/FiltersMenu/tests/index.test.js
@@ -1,9 +1,9 @@
 // import React from 'react'
 // import { shallow } from 'enzyme'
 
-// import NaviMenu from '../index'
+// import NaviCatalog from '../index'
 
-describe('TODO <NaviMenu />', () => {
+describe('TODO <NaviCatalog />', () => {
   it('Expect to have unit tests specified', () => {
     expect(true).toEqual(true)
   })
