+++ src/components/NaviCatalog/styles/header.js
