+++ src/components/NaviCatalog/styles/menu_list/pin_number.js
