+++ src/components/NaviCatalog/styles/pin_number.js
